Change RDF representation of compound subjects to get rid of bnodes
Until now we've used the MADS vocabulary to represent compound subjects
("subject strings"). In which the components of the compound subjects
are ordered using RDF lists, which unfortunately use blank nodes.

The problem with blank nodes is not that they are impossible to work
with, but with every new tool there are new problems and it just takes
too much time to work around all the problems. Latest in the line was
problems getting the components included correctly in the Skosmos REST
API responses.

Since we don't have a plan to provide a full MADS representation of the
vocabulary at the moment, we can use our own component attribute instead.
At the moment, we also don't have any application that requires an ordered
component list, so we can do with unordered sets. At least for now.